requirements.txt
asn1crypto==0.24.0
baidu-aip==2.2.7.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11.7
django-crispy-forms==1.7.2
idna==2.7
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
